+++ mandatory/src/ServerGetters.cpp
@@ -1,7 +1,7 @@
 #include "../inc/Server.hpp"
 
 
-size_t Server::getMaxClientBodySize() { return this->maxBodySize; }
+long long Server::getMaxClientBodySize() { return this->maxBodySize; }
 std::string Server::getMaxClientBodySizeStr() { return this->maxBodySizeStr; }
 bool Server::getIsDefault() { return this->isDefault; }
 std::string Server::getHost() { return this->Host; }
